pyproject.toml
[tool.poetry]
name = "Andrka-gendiff"
version = "0.8.0"
description = "Generate differences"
authors = ["Andrka <postbox4mail@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.0"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

